feat: Criação do recurso para download de paginas no formato html e implementação de processo em lote
